requirements.txt
astroid==3.3.11
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dill==0.4.0
Flask==2.3.2
Flask-Mail==0.10.0
Flask-WTF==1.2.2
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
platformdirs==4.3.8
pylint==3.3.7
pylint-flask==0.6
pylint-plugin-utils==0.9.0
python-dotenv==1.0.0
requests==2.32.4
tomlkit==0.13.3
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
Flask==2.3.2
python-dotenv==1.0.0